Put FilterRules in a separate file.
