+++ Assets/_Core/Scripts/Internal/EntitySystem/FilterRules.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: f64ab5c99924d4c4c9fab9f926e5e4eb
+timeCreated: 1543142507
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
